Remove YACReaderComicInfoPair metatype
When doing modularized builds, the definition of YACReaderComicInfoPair as
a metatype can lead to redefinition of Qt internal metatypes at the
moc stage. This is currently not happening in qmake, but is a problem for
our CMake migartion.

As we don't really use this type, it is safe to remove.